(SERVER-1494) Add smoke test
This commit adds a dependencies section that includes clojure.
This won't be inherited by child projects, but having it here
allows us to run `lein test` to just ensure that the project
is parseable.  It also inherits the clojure version from the
`managed-dependencies` section, so we get some basic smoke
testing of the inheritance.