(maint) Bump clj-i18n to 0.7.1
This commit picks up changes to clj-i18n that removes the use and generation of English bundles if an en.po is not present. Opting instead to fallback to the POT as intended. This work also fixes a bug surfacing GNU gettext treatment of empty strings. Prior to this commit, empty strings would display the header of the POT file. Now an externalized empty string displays an empty string.