(SERVER-1630) Bump snakeyaml version to 1.18
This commit bumps the org.yaml/snakeyaml version from 1.13 to 1.18, for
compatibility with the version used in the latest JRuby, 9.1.8.0.  Per
issue #323 in the snakeyaml bitbucket project, 1.18 does have some minor
breaking changes (method signature changes).  None of the changes
appear to be exposed / utilized from clj-yaml.  The next clj-parent
release which has this bump should probably be at least a Y version,
though, for visibility.