PE-23159: bumping rbac client version for consumer change
This commit bumps the rbac-client version for the project to include the new method added to the consumer service, which allows service calls for permitted instances with token authentication.

The changes also bump the clj-parent version with a minor increase, as the clj-rbac-client changeset is a breaking change to rbac because it updates the RbacConsumerService protocol.