(PDB-3942) Raise liberator dependency to java 9 compatible
Liberator can be made Java 9/10 compatible with the appropriate flags,
but without them, version 0.12 is not 9/10 compatible. Also there will
be no flag available to make liberator 0.12 compatible with Java 11.

This changes liberators version to the minimum required for 9/10
compatiblility without additional flags and the minimum required for
Java 11 compatibility.