(maint) update clj-http-client to 1.0.0
clj-http-client fixes an issue were malformed content-type headers
could result in a blank response.  It also adds two configuration
values that can be used for persistent clients to allow the total
number of connections allowed through the client, as well as the
total number of connections per route through the client.