(maint) Update dynapath to 1.0.0
This was updated in kitchensink when doing the Java 9+ work but because
it wasn't updated in clj-parent it doesn't look like most projects were
actually using the dependency at that version.

The previous version of dynapath (0.2.4) would decide whether or not it
could use the APIs exposed in Java 8 and if not only use the ones
exposed in Java 9+. It would make that decision at compile time however
so projects AOTed on Java 8 could not be ran on new Java versions. 0.2.5
moved the check to runtime rather than compile time, and 1.0.0 removes
the check completely and only extends the classes exposed in 9+.