(maint) Replace me.raynes/fs with clj-commons/fs
me.raynes/fs is unmaintained, though the project has been forked and
taken over by the clj-commons community. The API has not changed but
they are reactively updating dependencies as needed.

When we upgraded apache's commons/compress to 1.18 we broke raynes'
fs because it dependend on compress bringing in xz. This is only
apparent when require me.raynes.fs.compression.

Because this exposes the same namespaces and vars as raynes' I've chosen
to replace raynes with it. Otherwise which project was loaded would be
unspecified.

This will require all consumers to update their dependencies from
`[me.raynes/fs]` to `[clj-commons/fs]` but should require zero code
changes.