(PDB-4579) fix connection tests
PostgreSQL JDBC driver has a bug where the isValid() call on a
connection does not respect timeouts. This causes our connection pool to
not detect a dead connection, which will eventually exhaust our
connection pool entirely. See pgjdbc/pgjdbc#1557

This is a backport to an older clj-parent version that is used in pdb
6.3.x