(PDB-5000) Update nippy to 3.1.1
To fix the CVE this has a potentially breaking change to the freeze and
thaw methods. If a class you need to thaw is not on the allowlist you
will need to bind a new set of allowed java classes to thaw. It would
look something like the example below.
```
(binding [nippy/*thaw-serializable-allowlist* #{"org.joda.time.DateTime"}]
  (nippy/thaw (Files/readAllBytes path)))
```

Upgrading to 2.15.3 would have also negated the CVE, but 2.15.0 had a
breaking change even though it was a y release. And by going to the 3.y
versions we can use the allowlist instead of the initial whitelist that
they added in 2.15.0 and later renamed.