(PE-33177) Update http-client to 2.0.0
This version changes the default TLS protocols to be 1.3 and 1.2, and removes 1.1 and 1.0.

We are bumping the Y due to the potentially breaking change of removing old TLS protocols from http-client defaults. This enables TLSv1.3 on HTTP clients without having those settings exposed yet in config files for each component.