(PE-33226) Use pgjdbc fork to fix Bouncy Castle FIPS issue
The PostgreSQL JDBC driver overrides the chooseClientAlias method in a way that is incompatible with Bouncy Castle. It causes BC to choose the wrong signing algorithm. This fork fixes the issue. A PR has been submitted to pgjdbc, and hopefully we can switch back once that is accepted.

Downstream projects will need to add https://artifactory.delivery.puppetlabs.net/artifactory/maven to their list of repositories in project.clj.