(maint) Update to modern dynapath to resolve classpath violations
Dynapath is a library that allows clojure and java code to introspect
the dynamically created clojure classpaths. The versions we depended on
supported Java versions =< 8 and worked in Java versions >= 9 but caused
a warning to be issued on startup. Updating to a modern version of
dynapath removes support for versions of Java < 9 but also resolves the
warning on Java 9 and higher.