(PE-37972) update jdbc-util to 1.4.3
This commit updates jdbc-util to 1.4.3, which brings in a new version
that was built against the dependencies on clj-parent 5.6.15 which has
and updated version of pgjdbc. While we were not shipping a vulnerable
version of pgjdbc since that was up-to-date in clj-parent, the dep in
jdbc-util was still triggering the mend scans.